remove the redundant group indicators from authoring fix some ts error related to onRemove prop fix some styling related to the fixed toolbar
